+++ internal/identity/application/usecase/user_find_usecase.go
@@ -39,7 +39,7 @@ func (uc *userFindUseCase) Execute(ctx context.Context, input UserFindInput) (Us
 	ctx, span := otel.Trace().StartSpan(ctx, "UserFindUseCase.Execute")
 	defer span.End()
 
-	userModel, err := uc.userRepo.FindByID(ctx, uint(input.UserID))
+	userModel, err := uc.userRepo.FindByID(ctx, input.UserID)
 	if err != nil {
 		message := "error finding user by id %d"
 		uc.logger.Error(message, "error", err, "id", input.UserID)
@@ -47,7 +47,7 @@ func (uc *userFindUseCase) Execute(ctx context.Context, input UserFindInput) (Us
 	}
 
 	output := UserFindOutput{
-		UserID:   uint64(userModel.ID()),
+		UserID:   userModel.ID(),
 		Name:     userModel.Name(),
 		Email:    userModel.Email(),
 		Password: userModel.PasswordHash(),
