+++ internal/identity/domain/repository/account_confirmation_repository.go
@@ -9,6 +9,6 @@ import (
 type AccountConfirmationRepository interface {
 	Create(ctx context.Context, confirmation model.AccountConfirmationModel) (model.AccountConfirmationModel, error)
 	FindByToken(ctx context.Context, token string) (model.AccountConfirmationModel, error)
-	DeleteById(ctx context.Context, id uint) error
-	FindByUserID(ctx context.Context, userID uint) (model.AccountConfirmationModel, error)
+	DeleteById(ctx context.Context, id uint64) error
+	FindByUserID(ctx context.Context, userID uint64) (model.AccountConfirmationModel, error)
 }
